v2.1: 修复编辑/排序/分页/AI模板，优化UI
- 修复购买记录编辑弹窗state参数丢失(records.js包装函数)
- 价格总览编辑弹窗增加品牌、规格数量、规格单位字段
- 购买记录表格支持点击表头排序(日期/物品/品牌/分类/单价等)
- 修复分页翻到0页或负数的问题
- AI提示词模板优化：名称拆分规则、规格提取到specQty/specUnit
- 渠道筛选只显示有数据的渠道
- 增删改数据后自动刷新筛选下拉框和统计
- 保存/删除成功后右上角toast提示
- 备注列支持自动换行
- 弹窗间距优化，备注不再被截断
- 规格数量支持小数(step=any, parseFloat)

+++ js/records-modal.js
@@ -161,7 +161,7 @@ export function savePurchase() {
   const channel = document.getElementById('item-channel')?.value.trim() || '其他';
   const date = dateEl.value;
   const notes = document.getElementById('item-notes')?.value.trim() || '';
-  const specQty = parseInt(document.getElementById('item-spec-qty')?.value) || null;
+  const specQty = parseFloat(document.getElementById('item-spec-qty')?.value) || null;
   const specUnit = document.getElementById('item-spec-unit')?.value.trim() || null;
   const isPromo = document.getElementById('item-is-promo')?.checked || false;
   const promoType = document.getElementById('item-promo-type')?.value || null;
